refactor(plugins): move the chat message helpers out of routes/utils
extract_query_text and collect_document_ids read one thing each out of a list of chat messages
and touch nothing else. Both are used by the route and by the stream processor, so living under
routes made one of those callers reach sideways into the other's package for a pure function.
They move to chat/messages.py, next to the schemas they read.

stream_out_of_scope_refusal goes to stream_processor.py, the module that owns every other SSE
event the chat stream emits, and takes the refusal copy and gettext with it.

What is left in routes/utils is the three things that belong there: the RAG block resolution the
route performs before handing messages to the model, the tool list it resolves off the request,
and the metadata parsing the conversations route uses to build a response. The module is 150
lines, down from 349 before this pair of PRs.
